+++ src/main/kotlin/com/application/controller/dto/response/ServiceResponse.kt
@@ -1,7 +1,7 @@
 package com.application.controller.dto.response
 
-import com.application.domain.entity.ServiceAnnouncement
 import com.application.domain.entity.Services
+import com.application.domain.objects.ServiceAnnouncement
 import java.math.BigDecimal
 import java.time.Instant
 
@@ -10,7 +10,7 @@ data class ServiceResponse(
     var announcement: ServiceAnnouncement,
     var name: String,
     var description: String,
-    var time: Long,
+    var time: Int,
     var value: BigDecimal,
     var status: String,
     var createdAt: Instant,
